requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generativeai
python-dotenv
jsonschema
